Make sure internal state is initialized when toggling network mode.
If the MobileDataButton toggles the network mode, NetworkModeButton
relies on internal state when deciding whether a switch is needed.
That state, however, is only initialized in the updateState() method,
which is not guaranteed to be called prior to toggling state via
MobileDataButton if the settings app was removed from memory before.

Change-Id: Icb5e27635e1a6e3bf5e1f4b1264b33c18b414670